#53 - Persistent Gear - Refactored variables
To increase consistency. I believe the frayed use of the
CTI_P_LastPurchase and OnKilledGear variables could be causing the
issues reported.

Most players save their loadout as a template and work from that, so the
fact that this somewhat breaks the "Resupply" button in the gear menu is
moot, as they can just 'resupply' from their template.

Needs more in-depth testing.