Button removed, prices adjusted, base decay
This update removes the workers and associated buttons (Thanks to
Magnet) and instantly creates the factories.  This also adjusts prices
on the upgrades for both factions as well as increase in price to tanks
and cost of factories.  Parameter settings were tweaked to accommodate
the new prices.  Destroyed factories will now decay at a faster rate
once they're destroyed (after 5min they will be gone fast so repair them
quickly). Prices to heavy armor was increased in an effort to encourage
players to use the lighter armored variants. FOB bug was found and
corrected.