Add parameter to keep gear on death
Add parameter to keep gear on death, rather than reverting to your last
purchased gear set. Note that this is now the default option for this
parameter, which may need to be changed.

Ported from Zerty 1.3.0.0

Tested on a local server. Seems to work just fine. I remember when we
enabled a similar option a while ago on the main server that there was
some occasional wierdness, perhaps involving dying in a vehicle. This
should be tested more before being put in use on the main server.