Fix for persistent gear not surviving HALO
You would loose your backpack and goggles when HALO jumping with
persistent gear enabled. Fixed. #92.

The CTI_P_LastPurchase variable is updated elsewhere in the code every
two seconds whenever the player is alive, if persistent gear is on. This
means that when HALO jumping, the HALO gear changes are saved as if the
player had changed them, and overwrite the player's saved gear. So we
save the currently saved gear when the jump begins and load the gear
when the jump ends, regardless of if the player is dead or alive.

Doing the smarts in this file avoids tangling the knowledge of HALO
jumping into other code files.