Change: Expand town award radius
The town money award radius has been expanded so that anyone inside the
town gets money, while the actual capture radius stays small. This
removes the aggrivation of having a noob cap the town while everyone
else was waiting outside the capture radius.

I've made the radius fairly large. I feel like if you are in the town,
you should get the money, and not need to carefully assemble into a
small space.