Fix for backpacks not buying as vehicle cargo
When you would purchase a backpack from the equipment screen as with a
vehicle as a target, it would show in the gear screen, but never show in
the vehicle inventory after purchase. This fixes that.

Apparently backpacks need to be added to a vehicle with a different BIS
function than other items.

Tested locally, seems to work.