Fix for purchasing typeless items for a vehicle
Certain items (like the default opfor vest) are in the game, but not in
the list of purchasable gear. First moving such an item to a vehicles
inventory, then purchasing more gear for that vehicle would result in
the purchase failing to give you gear, since it crashed on looking up
the item's type.

Fixed.