Fix for fast capping of resistance towns
Resistance towns can not longer be fast capped. The problem had been
that with a tight town capture radius, the resistance were rarely inside
it, allowing the players to take the towns for free. I've changed the
way capturing towns works so that resistance units block town capture
anywhere in the town.

This fix does nothing for fast capping Opfor/Bluefor towns, but that's
for another day, and doesn't really matter because you just drive a tank
into town and clear it anyway.